Remove the old device API references in the Lightning repo and test suite (#1057)
**Context:**

Lightning simulators follow the new device API:
[lightning_newAPI_base.py](https://github.com/PennyLaneAI/pennylane-lightning/blob/master/pennylane_lightning/core/lightning_newAPI_base.py)
while we still keep references to the old device design and class
[lightning_base.py](https://github.com/PennyLaneAI/pennylane-lightning/blob/master/pennylane_lightning/core/lightning_base.py)
across our Python source code.

**Description of the Change:**
* Remove the old device API references in the Lightning repo
* Remove all obscured integration and unit tests related to the old
device API in the Lightning repository


**Benefits:**

**Possible Drawbacks:**

**Related GitHub Issues:**

[sc-83271] [sc-76160]

---------
