+++ tests/lightning_base/test_measurements_class.py
@@ -32,13 +32,6 @@ from pennylane.devices import DefaultQubit
 from pennylane.measurements import VarianceMP
 from scipy.sparse import csr_matrix, random_array
 
-if not LightningDevice._new_API:
-    pytest.skip(
-        "Exclusive tests for new API devices. Skipping.",
-        allow_module_level=True,
-    )
-
-
 if not LightningDevice._CPP_BINARY_AVAILABLE:
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)
