+++ tests/lightning_base/test_no_binaries.py
