+++ tests/lightning_base/test_simulate_method.py
@@ -22,12 +22,6 @@ import pytest
 from conftest import PHI, THETA, LightningDevice, device_name  # tested device
 from pennylane.devices import DefaultExecutionConfig, DefaultQubit
 
-if not LightningDevice._new_API:
-    pytest.skip(
-        "Exclusive tests for new API devices. Skipping.",
-        allow_module_level=True,
-    )
-
 if device_name == "lightning.tensor":
     pytest.skip("Skipping tests for the LightningTensor class.", allow_module_level=True)
