+++ tests/lightning_base/test_state_vector_class.py
@@ -36,12 +36,6 @@ if device_name == "lightning.gpu":
 if device_name == "lightning.tensor":
     pytest.skip("Skipping tests for the LightningTensor class.", allow_module_level=True)
 
-if not LightningDevice._new_API:
-    pytest.skip(
-        "Exclusive tests for new API devices. Skipping.",
-        allow_module_level=True,
-    )
-
 
 if not LightningDevice._CPP_BINARY_AVAILABLE:
     pytest.skip("No binary module found. Skipping.", allow_module_level=True)
